+++ Firmware/Sloth-Maple-Mini-Arduino-Gus/Motor.h
@@ -6,9 +6,9 @@
 class Motor {
     public:
         Motor(uint8_t pinPwm, uint8_t in1, uint8_t in2);
-        void speed(int speed);
+        void speed(float speed);
         void coast();
-        void brake();     
+        void brake();
     protected:
         uint8_t _pinPwm;
         uint8_t _in1;
